fix: address review — relational subgraph, hygiene counts, robustness
serialize(node_ids=): also keep 1-hop relation neighbours so a relational hit (what's on the table) still serves the table anchor + the on edge (was dropped — only hierarchy ancestors kept). Cycle-guard emit(); report the effective root in the header. bench.decay_forget: count nodes actually removed by forget() (not calls) + add nodes_after. Cast Replica instance ids to int64 + lock the np.where row/col invariant. Tests + recall_at_k doc caveat.

+++ src/spatialmem/datasets/__init__.py
@@ -142,7 +142,7 @@ def gt_detections_from_frame(
     """
     fx, fy, cx, cy = intrinsics
     depth = np.asarray(depth, dtype=np.float64)
-    instance = np.asarray(instance)
+    instance = np.asarray(instance).astype(np.int64)
     p = np.asarray(pose, dtype=np.float64)
     rot, trans = p[:3, :3], p[:3, 3]
     dets: list[Detection] = []
@@ -150,7 +150,7 @@ def gt_detections_from_frame(
         if iid in drop_ids:
             continue
         sel = (instance == iid) & (depth > 0)
-        vs, us = np.where(sel)
+        vs, us = np.where(sel)  # np.where -> (rows=v, cols=u); us drives x, vs drives y
         if vs.size < min_pixels:
             continue
         d = depth[vs, us]
